Allows both bosons and fermions
The earlier study on Fermions (i.e. no violations of Pauli Exclusion Principle) was defective (and resulted in the total number of particles decaying to zero). This current version still needs verification that there are no  possible PEP-compliant inputs to a node for which the output flows would have to violate PEP (I play to prove that simply by machine verification of all possible scenarios), but several thousands of iterations have produced no violations of PEP so this is a lot better than the colossally bad (as in exploding to numerical overflow in just a couple of dozen time-steps) previous versions. They're based on just doing standard ModPool crunching of any PEP-compliant inputs and using the heat bath to shave off any deformations and to allow them to cancel out, more or less.